extracted components from the account page
